+++ persistance_layer/query.js
@@ -2,35 +2,58 @@ import connect2 from './singlton_provider2.js';
 import connect1 from './singlton_provider1.js';
 import connect_main from './singlton_main_db.js';
 
-
-class Builder {
-  constructor() {
-    this.connection
+class Query{
+  constructor(){
+    this.connection = undefined
+    this.string_query  = undefined
   }
 
+  set_connection(connection){
+    this.connection = connection
+    return this
+  }
 
-  async select_filter(lower, uper){ //if prov 1
-    this.connection = connect1
-    const result = await this.connection.query(`select*from getting_list_descriptin_by_price('${lower}', '${uper}');`)
-    console.log("builder \n " + result.rows)
-    return result.rows
+  set_string_query(str_query){
+    this.string_query = str_query
+    return this
+  }
+}
+
+class Query_Constructor{
+  constructor(){
+    this.query = new Query
+  }
+  
+  async select_filter(lower, uper){ //if prov 1 
+    this.query = this.query
+                          .set_connection(connect1)
+                          .set_string_query(`select*from getting_list_descriptin_by_price('${lower}', '${uper}');`)
+    return await this.send_query()
   }
 
   async select(){ //if main db
-    this.connection = connect_main
-    const result = await this.connection.query(`select*from getting_list_description();`)
-    return result.rows
+    this.query = this.query
+                          .set_connection(connect_main)
+                          .set_string_query(`select*from getting_list_description();`)
+    return await this.send_query()
   }
 
   async select_list(){ //if prov 2
-    this.connection = connect2
-    const  result = await this.connection.query(`select*from getting_price_list_real_estate_by_name();`);
-    return result.rows;
+    this.query = this.query
+                          .set_connection(connect2)
+                          .set_string_query(`select*from getting_price_list_real_estate_by_name();`)
+    return await this.send_query()
   }
 
   async select_details(id){ //if prov 2
-    this.connection = connect2
-    const result  = await this.connection.query(`select*from getting_detail_by_id_real_estate(${id});`)
+    this.query = this.query
+                          .set_connection(connect2)
+                          .set_string_query(`select*from getting_detail_by_id_real_estate(${id});`)
+    return await this.send_query()
+  }
+
+  async send_query(){
+    const result = await this.query.connection.query(this.query.string_query)
     return result.rows
   }
 
@@ -48,4 +71,4 @@ class Builder {
 										 descr text, price float, 
 										 area_m float, type_real_estate varchar (255) */
 
-export {Builder};
+export default Query_Constructor;
